Remove URL from PDFs. use FullName of agent.
Martin removed URLs from PDF and HTML prints in OTRS 3.3. This
commit applies this also to ITSM.

Marco Buchholz added a configuration option to influence how an
agent name is rendered: 'firstname lastname', 'lastname firstname'
or variations of that including a login. This is for consistency
reasons. This commit makes ITSM use this setting.